Fix order quantity calculation to use integer conversion for minimum order quantity

+++ arbitrage_telegram_alarm_bot/producer.py
@@ -111,7 +111,7 @@ async def order_handler(data):
             symbol=data.get('ticker'),
             side='bid' if data.get('candle_type') == '-' else 'ask',
             order_type='market',
-            qty=float(minOrderQty)*10,
+            qty=int(minOrderQty*10),
             tp=tp,
             sl=sl
         )
